Update WAF security import to pull in existing data
Currently, when importing a WAF security rule, only the site and rule
id. The import logic needs to be updated to read the existing rule from
incapsula to prevent any needed changes upon planning. After importing,
all the fields are left as null. When updating the import logic to
perform a read after populating the proper IDs, all the fields are
populated correctly now upon import.