feat: Allow name_id to be specifified in ABP website groups to discriminate
We need something unique specified by the user for each item in a terraform array, otherwise any removal or additions in the middle of the array will shuffle
all elements after the change in such a way to we end up re-creating all the websites there. Using name is the most obvious way to have a unique identifier however
name is allowed to be non-unique so adding `name_id` as a secondary (but arbitrary) identifier lets users only specify name in the common case of unique names and opt-in
to specifying `name_id` when `name` shouldn't be unique.